channel_the_elements() on sand should use *Fire* skill level
It seems very weird that the sand-based variant of
channel_the_elements uses a *spell* which is only related via the "Fire"
skill to determine whether to use hellfire or regular fire. This change
ensures that it is consistent with the rest of channel_the_elements and
ensures that we're in line with the documentation of Geomancy.