Remove easy_{disarm,tunnel,open}
We behave is if the options (in the old code) were set to their
defaults, *except* for easy_disarm which is rather dangerous.
Because of the dangerousness we behave as if easy_disarm==FALSE,
except for picking the "obvious" choice if there is one when
using the (D)isarm command.