Do not return NULL when Esc is hit in Automizer
Formerly many functions would return NULL when this happened, leading to
an assert() failure and immediate crash further up the call stack. This
patch makes them procede as normal, allowing a default or even empty
automizer rule to be added. This is not necessarily good, but is
preferable to crashing the game.