Do not consume charges for wish if input is 'invalid'
Previously making a wish with no outcome, i.e. no summon, no item,
etc. would use a charge. Given the extreme pickiness of the code that
was extremely bad UX, so this changes the code to NOT use a charge.
It's still pretty bad UX, but at least it doesn't consume a very
finite resource, and the player can retry.