Fix loading macros, see #51
The old code used to compare F1] to F1 and would therefor always fail. I am not an expert c coder and I am fairly certain that there should be a better way of doing what this commit does. I also think that most of the text_to_ascii and ascii_to_text code is in dire need of a cleanup.

At least with this change I am able to play the game without having to
redo my macros at the start of every session.